+++ .gitignore
@@ -15,7 +15,6 @@ dist-docs
 stats.html
 
 # Editor directories and files
-.vscode/*
 !.vscode/extensions.json
 .idea
 .DS_Store
