+++ src/filters/index.ts
@@ -1,0 +1,4 @@
+export { booleanFilterFn } from './booleanFilter';
+export { dateFilterFn } from './dateFilter';
+export { numberFilterFn } from './numberFilter';
+export { stringFilterFn, highlightFilterValue } from './stringFilter';
