+++ tsconfig.json
@@ -15,7 +15,6 @@
         "isolatedModules": true,
         "jsx": "react-jsx",
         "declaration": true,
-        "emitDeclarationOnly": true,
         "outDir": "dist"
     },
     "include": ["src"],
